+++ src/context/useUIManager.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect, useCallback, useRef } from "react";
 import { EDIT_MODE, RUNTIME_MODE, type Mode } from "@src/constants/constants";
 import { useWidgetManager } from "./useWidgetManager";
-import type { Widget } from "@src/types/widgets";
+import type { ExportedWidget, Widget } from "@src/types/widgets";
 import useEpicsWS from "./useEpicsWS";
 import {
   authService,
@@ -14,6 +14,7 @@ import { notifyUser } from "@src/services/Notifications/Notification";
 import {
   getAllDeployedReposTree,
   getAllReposTree,
+  updateStagingRepoFile,
   type RepoTreeInfo,
   type User,
 } from "@src/services/APIClient";
@@ -29,9 +30,14 @@ export interface SelectedPathInfo {
 export default function useUIManager(
   ws: ReturnType<typeof useEpicsWS>,
   setSelectedWidgetIDs: ReturnType<typeof useWidgetManager>["setSelectedWidgetIDs"],
+  editorWidgets: ReturnType<typeof useWidgetManager>["editorWidgets"],
+  formatWdgToExport: ReturnType<typeof useWidgetManager>["formatWdgToExport"],
   fileLoadedTrig: ReturnType<typeof useWidgetManager>["fileLoadedTrig"],
 ) {
   const lastFileLoadedTrig = useRef(0);
+  const hasFileChanged = useRef(true);
+  const lastSavedRef = useRef<ExportedWidget[] | null>(null);
+  const saveTimeoutRef = useRef<number | null>(null);
   const [releaseShortcuts, setReleaseShortcuts] = useState(false);
   const [wdgPickerOpen, setWdgPickerOpen] = useState(false);
   const [pickedWidget, setPickedWidget] = useState<Widget | null>(null);
@@ -69,6 +75,7 @@ export default function useUIManager(
       ws.startNewSession();
       lastFileLoadedTrig.current = fileLoadedTrig;
     }
+    hasFileChanged.current = true;
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [fileLoadedTrig]);
 
@@ -141,6 +148,52 @@ export default function useUIManager(
       }
     };
   }, [inEditMode, ws]);
+
+  // throttle file update to backend
+  useEffect(() => {
+    if (!isDeveloper || !inEditMode) return;
+    if (!selectedFile?.repo_id || !selectedFile.path) return;
+    // Skip the first render after selecting a new file
+    if (hasFileChanged.current) {
+      hasFileChanged.current = false;
+      return;
+    }
+    const exportable = editorWidgets.map(formatWdgToExport);
+    // Skip if content didn't change
+    if (lastSavedRef.current === exportable) return;
+    const serialized = JSON.stringify(exportable, null, 2);
+
+    // debounce
+    if (saveTimeoutRef.current) {
+      clearTimeout(saveTimeoutRef.current);
+    }
+
+    const updateFileContent = async () => {
+      try {
+        const updtd = await updateStagingRepoFile({
+          path: { repo_id: selectedFile.repo_id },
+          query: { path: selectedFile.path },
+          body: { content: serialized },
+        }).then((r) => r.data);
+
+        setReposTreeInfo((prev) => {
+          if (!prev) return prev;
+          return prev.map((r) => (r.id === updtd.id ? updtd : r));
+        });
+        lastSavedRef.current = exportable;
+      } catch (err) {
+        notifyUser(`Failed to save file: ${err as string}`, "error");
+      }
+    };
+
+    saveTimeoutRef.current = window.setTimeout(() => void updateFileContent(), 500);
+
+    return () => {
+      if (saveTimeoutRef.current) {
+        clearTimeout(saveTimeoutRef.current);
+      }
+    };
+  }, [editorWidgets, selectedFile, isDeveloper, inEditMode, formatWdgToExport, setReposTreeInfo]);
 
   return {
     releaseShortcuts,
